feat: add `auto-config` and `build` inputs.
`auto-config` allows users to specify if `lean-action` should
use the Lake workspace to automatically decide which CI features to run.

`build` allows users to specify if `lean-action` runs `lake build`.

By default, `auto-config: true`.

The `test` and `build` (and soon `lint`, see #46) inputs
allow users to override the automatically configured behavior
or configure `lean-action` when `auto-config: false`.

`auto-config: true` is close to the previous default behavior,
however there is a difference in the outcome of the `lake test` step.
When users set `test: true` manually, `lean-action` must find tests
with `lake check-test` and run `lake test` or it will fail
(this was the previous behavior). However with `auto-config: true`,
if `lake check-test` fails, `lean-action` will not run `lake test`
and this won't cause `lean-action` to fail.