feat: option for running lake exe mk_all --check (#129)
This command checks that all files in the default target's directory are
imported in the main file. In repositories such as FLT, this check is
run before other build steps, to ensure every relevant file is built,
but it must be run after Elan has been initialized. So to use
`lean-action` in FLT, we would like to incorporate the mk_all check as a
step.

---------

Co-authored-by: Austin Letson <waustinletson@gmail.com>